requirements.txt
faunadb
python-dotenv
pathlib